Refactored FarmLand
