+++ src/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_executable(barren-land-calculator
     main.cpp
+    FarmLand.cpp
     InputParser.cpp
     Point.cpp
     Rectangle.cpp
